Refactored counterfactual tests and extracted common fixtures
