+++ teacher/explanation/tests/test_factual.py
@@ -1,40 +1,5 @@
-import pytest
 from teacher.tree import Rule
 from teacher.explanation import m_factual, c_factual, mr_factual, FID3_factual
-
-
-@pytest.fixture
-def mock_fired_rules():
-
-    instance = {
-        'color': {
-            'low': 0.5,
-            'high': 0.5
-        },
-        'bitterness': {
-            'very high': 0.8,
-            'extremely high': 0.2
-        },
-        'strength': {
-            'high': 0.47,
-            'very high': 0.53
-        }
-    }
-
-    class_val = 'IPA'
-
-    rule_one = Rule([('color', 'low'), ('strength', 'high')], 'IPA', 0.93)
-    rule_two = Rule([('color', 'high'), ('strength', 'high')], 'IPA', 0.65)
-    rule_three = Rule([('color', 'low'), ('strength', 'very high')], 'IPA', 0.44)
-    rule_four = Rule([('color', 'high'), ('strength', 'very high')], 'IPA', 0.12)
-    rule_five = Rule([('color', 'high'), ('strength', 'very high')], 'Barleywine', 0.88)
-    rule_six = Rule([('color', 'low'), ('strength', 'very high')], 'Barleywine', 0.28)
-    rule_seven = Rule([('color', 'high'), ('strength', 'high')], 'Barleywine', 0.35)
-    rule_eight = Rule([('color', 'low'), ('strength', 'high')], 'Barleywine', 0.07)
-
-    rule_list = [rule_one, rule_two, rule_three, rule_four, rule_five, rule_six, rule_seven, rule_eight]
-
-    return instance, rule_list, class_val
 
 
 def test_m_factual(mock_fired_rules):
@@ -65,7 +30,6 @@ def test_mr_factual(mock_fired_rules):
     rule_one = Rule([('color', 'low'), ('strength', 'high')], 'IPA', 0.93)
     rule_two = Rule([('color', 'high'), ('strength', 'high')], 'IPA', 0.65)
     rule_three = Rule([('color', 'low'), ('strength', 'very high')], 'IPA', 0.44)
-
     expected_factual = [rule_one, rule_two, rule_three]
     assert mr_factual(instance, rule_list, class_val) == expected_factual
